dom/serviceworkers/test/test_importscript_mixedcontent.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1198078 - test that we respect mixed content blocking in importScript() inside service workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<iframe></iframe>
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var iframe;
  function runTest() {
    iframe = document.querySelector("iframe");
    iframe.src = "https://example.com/tests/dom/serviceworkers/test/fetch/importscript-mixedcontent/register.html";
    var ios;
    window.onmessage = function(e) {
      if (e.data.status == "ok") {
        ok(e.data.result, e.data.message);
      } else if (e.data.status == "registrationdone") {
        iframe.src = "https://example.com/tests/dom/serviceworkers/test/fetch/importscript-mixedcontent/index.html";
      } else if (e.data.status == "done") {
        is(e.data.data, "good", "Mixed content blocking should work correctly for service workers");
        iframe.src = "https://example.com/tests/dom/serviceworkers/test/fetch/importscript-mixedcontent/unregister.html";
      } else if (e.data.status == "unregistrationdone") {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        SimpleTest.finish();
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
      ["security.mixed_content.block_active_content", false],
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
